BackEnd/requirements.txt
###### Requirements ######
aiohttp
mysql.connector
dataclasses
typing